requirements.txt
torch==2.2.2
torchvision==0.17.2
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.2.0
tqdm==4.66.2
sentencepiece==0.1.99
transformers==4.38.2
git+https://github.com/openai/CLIP.git